updated datatypes
added float32 & float64

+++ Dexterous-Engine/src/dex/Renderer/Mesh/Mesh.hpp
@@ -12,7 +12,7 @@ namespace dex
 {
 	namespace Mesh
 	{
-		enum class Type : uint8_t
+		enum class Type : uint32
 		{
 			BASE,
 			INTERFACE,
